Add '-c' option to generate_certbot to choose which types of certificates to request
The script will request all certificates by default, or the value of the
environment variable ALTERNC_REQUEST_CERTIFICATES if is set.

Changes were made to the install.d script and the cronjob to support this. The
cron job will request all certificates explicitly. The install.d script will
default to all if the variable isn't set in local.sh, but will override the
environment variable set if the flag '--all' is present in the arguments.